refactor(web-app-ai-multi-doc-synthesizer): use dispatchModal, fix type errors, clean up
Architecture fix:
- defineWebApplication does not support rootComponent; remove it
- Adopt useModals().dispatchModal() pattern (same as ai-sensitive-data-scanner)
  so the synthesis UI is opened as a proper oc-modal dialog
- Replace SynthesisOverlay.vue (custom backdrop + Teleport) with SynthesisPanel.vue
  (modal body content only, no backdrop/teleport — modal handles that)
- Remove App.vue and state.ts (global ref state no longer needed with dispatchModal)
- index.ts now passes resources + llmConfig as customComponentAttrs

Type errors fixed:
- Mock return values in tests now use `as any` (same as useChat.spec.ts reference)
- File-support spec unchanged and still passes

Tests: 22/22 passing, check:types clean
